Add unit test for parameters and fix bug
