+++ tests/helpers/options.js
@@ -1,0 +1,10 @@
+module.exports = {
+  sourceBranch: "master",
+  targetBranch: "release",
+  upstream: "origin",
+  tagPattern: "",
+  prefix: false,
+  silent: true,
+  verbose: false,
+  dryrun: false
+};
